Created an helper to fix React act() warnings
